石川1108-1332

+++ sql/CREATE_buydetails.sql
@@ -1,8 +1,8 @@
 CREATE TABLE buydetails(
-    buy_id          INT(20) NOT NULL,
-    buy_detailsid   INT(20)     NOT NULL,
-    shohin_id       INT(10) NOT NULL,
-    shohin_count    INT(20)     NOT NULL,
+    buy_id          INT     NOT NULL,
+    buy_detailsid   INT     NOT NULL,
+    shohin_id       BIGINT  NOT NULL,
+    shohin_count    INT     NOT NULL,
     PRIMARY KEY(buy_id,buy_detailsid),
     FOREIGN KEY(buy_id)
     REFERENCES buy(buy_id),
