+++ sql/CREATE_carts.sql
@@ -1,7 +1,7 @@
 CREATE TABLE carts(
     mem_mail        VARCHAR(200)    NOT NULL,
-    shohin_id       INT(10) NOT NULL,
-    shohin_count    INT(20)     NOT NULL,
+    shohin_id       BIGINT          NOT NULL,
+    shohin_count    INT             NOT NULL,
     PRIMARY KEY(mem_mail,shohin_id),
     FOREIGN KEY(mem_mail)
     REFERENCES members(mem_mail),
